+++ tilings/__init__.py
@@ -2,6 +2,6 @@ from tilings.assumptions import TrackingAssumption
 from tilings.griddedperm import GriddedPerm
 from tilings.tiling import Tiling
 
-__version__ = "2.3.1"
+__version__ = "2.5.0"
 
 __all__ = ["GriddedPerm", "Tiling", "TrackingAssumption"]