+++ tilings/algorithms/__init__.py
@@ -3,6 +3,7 @@ from .factor import Factor, FactorWithInterleaving, FactorWithMonotoneInterleavi
 from .fusion import ComponentFusion, Fusion
 from .gridded_perm_generation import GriddedPermsOnTiling
 from .gridded_perm_reduction import GriddedPermReduction
+from .guess_obstructions import guess_obstructions
 from .minimal_gridded_perms import MinimalGriddedPerms
 from .obstruction_inferral import (
     AllObstructionInferral,
@@ -12,6 +13,7 @@ from .obstruction_inferral import (
 from .obstruction_transitivity import ObstructionTransitivity
 from .requirement_placement import RequirementPlacement
 from .row_col_separation import RowColSeparation
+from .sliding import Sliding
 from .subclass_verification import SubclassVerificationAlgorithm
 
 __all__ = [
@@ -33,4 +35,6 @@ __all__ = [
     "RequirementPlacement",
     "RowColSeparation",
     "SubclassVerificationAlgorithm",
+    "guess_obstructions",
+    "Sliding",
 ]